IAS 1.3 : > Auto Member Inviter completed ( check if debugging is required )
