+++ main.py
@@ -1,9 +1,9 @@
 from flask import Flask
 from flask_cors import CORS
-from shared import app, config
+from shared import app, config, invitation_sender
 from modules.public.api import public_blueprint
 from modules.points.api import points_blueprint
-from modules.users.api import InvitationSender, create_users_blueprint  # Import the necessary components
+from modules.users.api import users_blueprint
 from modules.utils.db import DBConnect
 
 # Initialize CORS with explicit origins allowed
@@ -13,17 +13,10 @@ CORS(app, resources={r"/*": {"origins": "http://127.0.0.1:3000/", }})
 db_connect = DBConnect()
 db_connect.check_and_create_tables()
 
-# Prompt for credentials before initializing InvitationSender
-username = input("Enter your ASU ID: ")
-password = input("Enter your ASU password: ")
-
-# Instantiate the InvitationSender class with credentials
-invitation_sender = InvitationSender(username, password)
-
 # Register Blueprints
 app.register_blueprint(public_blueprint, url_prefix="/")
 app.register_blueprint(points_blueprint, url_prefix="/points")
-app.register_blueprint(create_users_blueprint(invitation_sender), url_prefix="/users")
+app.register_blueprint(users_blueprint, url_prefix="/users")
 
 if __name__ == "__main__":
     # Login once when the application starts
